+++ core/constants.py
@@ -11,7 +11,7 @@ import numpy as np
 
 # ── Metadatos de la aplicación ────────────────────────────────────────────────
 APP_NAME = "Fitbauer"
-APP_VERSION = "4.11.2"
+APP_VERSION = "4.11.3"
 APP_AUTHOR = "Jorge Sánchez Marcos, Nieves Menéndez González"
 APP_DEPARTMENT = "Departamento de Química Física · UAM"